refactor(onboard): extract docker gateway runtime helpers
